# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Ephemeral VM Lab API
  version: 1.0.0
  description: >
    REST API for the orion-complex Ephemeral VM Lab.
    Manages disposable VM environments across libvirt (Linux) and
    Apple Virtualization.framework (macOS) providers.
servers:
  - url: http://127.0.0.1:3000
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: string }
        provider: { type: string }
        provider_subject: { type: string }
        email: { type: string }
        email_domain: { type: string }
        display_name: { type: string }
        role: { type: string, enum: [user, admin] }
        disabled: { type: integer }
        created_at: { type: integer }
...
    Environment:
      type: object
      properties:
        id: { type: string }
        image_id: { type: string }
        owner_user_id: { type: string }
        node_id: { type: string }
        provider: { type: string }
        guest_os: { type: string }
        guest_arch: { type: string }
        state:
          type: string
          enum: [creating, running, suspending, suspended, resuming, rebooting, migrating, destroying, failed]
        created_at: { type: integer }
        expires_at: { type: integer, nullable: true }
        vcpus: { type: integer }
        memory_bytes: { type: integer }
        disk_bytes: { type: integer }
    Snapshot:
      type: object
      properties:
        id: { type: string }
        env_id: { type: string }
        name: { type: string, nullable: true }
        created_at: { type: integer }
    Task:
      type: object
      properties:
        id: { type: string }
        kind: { type: string }
        state: { type: string, enum: [pending, running, completed, failed] }
        created_at: { type: integer }
        error_message: { type: string, nullable: true }
        completed_at: { type: integer, nullable: true }
        env_id: { type: string, nullable: true }
...
    UserSshKey:
      type: object
      properties:
        id: { type: string }
        user_id: { type: string }
        public_key: { type: string }
        fingerprint: { type: string }
        created_at: { type: integer }
...
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, maximum: 200 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, default: 0 }
security:
  - bearerAuth: []
paths:
...
  /v1/users:
    get:
      summary: List users (admin only)
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
  /v1/users/{user_id}:
    get:
      summary: Get user details (admin only)
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
  /v1/users/{user_id}/role:
    put:
      summary: Update user role (admin only)
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [user, admin] }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
  /v1/users/{user_id}/disabled:
    put:
      summary: Enable/disable user (admin only)
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [disabled]
              properties:
                disabled: { type: boolean }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
  /v1/users/{user_id}/ssh-keys:
    get:
      summary: List SSH keys for a user (provisioning endpoint)
      parameters:
        - name: user_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UserSshKey' }

  # ── SSH Keys ───────────────────────────────────────────────────────
  /v1/ssh-keys:
    get:
      summary: List current user's SSH keys
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/UserSshKey' }
    post:
      summary: Add SSH key
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [public_key]
              properties:
                public_key: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSshKey' }
  /v1/ssh-keys/{key_id}:
    delete:
      summary: Delete SSH key
      parameters:
        - name: key_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Key deleted

  # ── Nodes ──────────────────────────────────────────────────────────
...
  /v1/environments/{env_id}/snapshots:
    get:
      summary: List snapshots for environment
      parameters:
        - name: env_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Snapshot' }
    post:
      summary: Create snapshot (requires running or suspended state)
      parameters:
        - name: env_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
      responses:
        '201':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Snapshot' }
  /v1/environments/{env_id}/snapshots/{snapshot_id}:
    delete:
      summary: Delete snapshot
      parameters:
        - name: env_id
          in: path
          required: true
          schema: { type: string }
        - name: snapshot_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: Snapshot deleted
  /v1/environments/{env_id}/snapshots/{snapshot_id}/restore:
    post:
      summary: Restore snapshot (requires running or suspended state)
      parameters:
        - name: env_id
          in: path
          required: true
          schema: { type: string }
        - name: snapshot_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Environment' }

  # ── Tasks ──────────────────────────────────────────────────────────
  /v1/tasks:
    get:
      summary: List all tasks
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }
  /v1/tasks/{task_id}:
    get:
      summary: Get task details
      parameters:
        - name: task_id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Task' }
  /v1/environments/{env_id}/tasks:
    get:
      summary: List tasks for a specific environment
      parameters:
        - name: env_id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Task' }

  # ── Events ─────────────────────────────────────────────────────────
...